#7 ♻️refactor: 엔티티 coin int로 변경

+++ src/main/java/com/sparta/devquiz/domain/coin/dto/request/SaveCoinRequest.java
@@ -8,6 +8,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class SaveCoinRequest {
 
-    private final long coins;
+    private final int coins;
     private final CoinContent coinContent;
 }
