+++ src/main/java/com/sparta/devquiz/domain/coin/dto/request/UseCoinRequest.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class UseCoinRequest {
 
-    private final long coins;
+    private final int coins;
     private final CoinContent coinContent;
 
 }
