+++ src/main/java/com/sparta/devquiz/domain/coin/enums/CoinContent.java
@@ -10,19 +10,19 @@ import java.util.function.Supplier;
 public enum CoinContent {
 
     // 문제
-    FIRST("SAVE", "FIRST", () -> 20L),
-    CORRECT("SAVE","CORRECT", () -> 10L),
-    FAIL("SAVE","FAIL", () -> 5L),
-    PASS("SAVE","PASS", () -> 0L),
-    USE("USE","USE", () -> 0L),
+    FIRST("SAVE", "FIRST", () -> 20),
+    CORRECT("SAVE","CORRECT", () -> 10),
+    FAIL("SAVE","FAIL", () -> 5),
+    PASS("SAVE","PASS", () -> 0),
+    USE("USE","USE", () -> 0),
 
 
     // 아이템
-    ITEM_CAT("USE", "CAT", () -> 25L),
-    ITEM_DOG("USE", "DOG", () -> 45L);
+    ITEM_CAT("USE", "CAT", () -> 25),
+    ITEM_DOG("USE", "DOG", () -> 45);
 
     private final String status;
     private final String content;
-    private final Supplier<Long> coinSupplier;
+    private final Supplier<Integer> coinSupplier;
 
 }
